Ambiguity and whitespace fixes in src/graphwidget.cpp

+++ src/graphwidget.cpp
@@ -255,16 +255,16 @@ void GraphWidget::animateRange(uint64_t _min, uint64_t _max)
 {
 	QPropertyAnimation* animation = new QPropertyAnimation(this, "minTime");
 	animation->setDuration(150);
-	animation->setStartValue(m_minTime);
-	animation->setEndValue(_min);
+	animation->setStartValue(qulonglong(m_minTime));
+	animation->setEndValue(qulonglong(_min));
 	animation->start();
 
 	connect(animation, SIGNAL(valueChanged(const QVariant&)), this, SLOT(zoomAnimEvent()));
 
 	animation = new QPropertyAnimation(this, "maxTime");
 	animation->setDuration(150);
-	animation->setStartValue(m_maxTime);
-	animation->setEndValue(_max);
+	animation->setStartValue(qulonglong(m_maxTime));
+	animation->setEndValue(qulonglong(_max));
 	animation->start();
 }
 
@@ -419,7 +419,7 @@ void GraphWidget::resizeEvent(QResizeEvent* _event)
 		m_grid->parentResized();
 
 	Q_FOREACH( GraphCurve* it, m_curves ) {
-		it->parentResized(); 
+		it->parentResized();
 	}
 
 	fitInView(m_scene->sceneRect());
@@ -493,7 +493,7 @@ void GraphWidget::mouseMovement(const QPoint& _position, Qt::MouseButtons _butto
 			break;
 		}
 	}
-	
+
 	if (rect().contains(pos) && m_LButtonDown && !m_inContextMenu)
 	{
 		uint64_t pL = mapPosToTime(m_dragStartPos.x());
@@ -507,9 +507,9 @@ void GraphWidget::mouseMovement(const QPoint& _position, Qt::MouseButtons _butto
 		float startTimeF = m_context->m_capture->getFloatTime(startTime);
 		float endTimeF = m_context->m_capture->getFloatTime(endTime);
 
-		QString ttip =	tr("Start time") + ": " + getTimeString(startTimeF) + "\n" + tr("End time") + ": " + getTimeString(endTimeF) + "\n" + 
-						tr("Duration") + ": " + getTimeString(endTimeF-startTimeF) + "\n" + tr("Usage at end") + ": " + m_locale.toString(entry.m_usage) + "\n" + 
-						tr("Live blocks") + ": " + m_locale.toString(entry.m_numLiveBlocks);
+		QString ttip =	tr("Start time") + ": " + getTimeString(startTimeF) + "\n" + tr("End time") + ": " + getTimeString(endTimeF) + "\n" +
+						tr("Duration") + ": " + getTimeString(endTimeF-startTimeF) + "\n" + tr("Usage at end") + ": " + m_locale.toString(qulonglong(entry.m_usage)) + "\n" +
+						tr("Live blocks") + ": " + m_locale.toString(qulonglong(entry.m_numLiveBlocks));
 		QToolTip::showText(gpt,ttip,this,QRect(pos,pos));
 	}
 	else
@@ -517,7 +517,7 @@ void GraphWidget::mouseMovement(const QPoint& _position, Qt::MouseButtons _butto
 		if (dontHideToolTip == false)
 			QToolTip::hideText();
 	}
-	
+
 	if ((_buttons & Qt::LeftButton) && m_LButtonDown && !m_isDragging)
 		m_isDragging = true;
 
@@ -571,7 +571,7 @@ void GraphWidget::mouseMovement(const QPoint& _position, Qt::MouseButtons _butto
 
 			newMaxTime = qMin(newMaxTime, (int64_t)m_context->m_capture->getMaxTime());
 			newMinTime = qMax(newMinTime, (int64_t)m_context->m_capture->getMinTime());
-			
+
 			m_minTime = newMinTime;
 			m_maxTime = newMaxTime;
 
@@ -586,10 +586,10 @@ uint64_t GraphWidget::currentPos()
 {
 	QRect drawRect = getDrawRect();
 	QPoint gpt = QCursor::pos();
-	
+
 	QPoint pos = mapFromGlobal(gpt);
 	QPointF pt = mapToScene(pos);
-	
+
 	if (drawRect.contains(pt.x(),pt.y()) && !m_LButtonDown && !m_inContextMenu)
 	{
 		uint64_t pL = mapPosToTime(pos.x());
@@ -599,8 +599,8 @@ uint64_t GraphWidget::currentPos()
 		QString timeStr = getTimeString(timeF);
 
 		QString ttip =	tr("Time") + ": " + timeStr + "\n" +
-						tr("Usage") + ": " + m_locale.toString(entry.m_usage) + "\n" +
-						tr("Live blocks") + ": " + m_locale.toString(entry.m_numLiveBlocks);
+						tr("Usage") + ": " + m_locale.toString(qulonglong(entry.m_usage)) + "\n" +
+						tr("Live blocks") + ": " + m_locale.toString(qulonglong(entry.m_numLiveBlocks));
 		QToolTip::showText(gpt,ttip,this,QRect(pos,pos));
 		return pL;
 	}
